contato service

+++ src/main/java/com/minsait/JPs/repository/ContatoRepository.java
@@ -3,5 +3,5 @@ package com.minsait.JPs.repository;
 import com.minsait.JPs.model.Contato;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContatoRepository extends JpaRepository<Contato,Integer>{
+public interface ContatoRepository extends JpaRepository<Contato,Long>{
 }
