+++ src/main/java/com/minsait/JPs/repository/PessoaRepository.java
@@ -3,5 +3,5 @@ package com.minsait.JPs.repository;
 import com.minsait.JPs.model.Pessoa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PessoaRepository extends JpaRepository<Pessoa,Integer> {
+public interface PessoaRepository extends JpaRepository<Pessoa,Long> {
 }
